Revised to work in a Windows 10 Universal webview, especially the window.alert() function.
Also added the condition to only execute the compat.js functionality, originally created by Microsoft, if you're in a Win 8.x webview, since it is not needed for running in a Win 10 webview. And included some comments providing details on how to detect specific Windows webviews as well as adding a simpler alternate window.alert() implementation for Windows that may or may not work, it is untested at this time. That alternate implemenation is commented out and can be ignored until time permits to determine if it has any value.